Move files so they work well with zeitwork autoloading
The Rails app's loader incorrectly recorded that AddressAssociations was a top-level constant:
  address_concern-2.1.0/app/models/concerns/address_associations.rb"=>[Object, :AddressAssociations]
when in fact it is nested under AddressConcern.

Renamed to lib/address_concern/address_associations.rb, for example, so Zeitwerk would detect the
namespace, AddressConcern::AddressAssociations.

By moving to lib, the parent Rails app should no longer try to auto-load those files, so we didn't
_need_ to follow that convention, but we also changed to use zeitwork _internally_ to load files in
this library (from lib), and it doesn't hurt to follow that convention.